+++ src/main/java/be/baur/sds/serialization/SDSParser.java
@@ -11,20 +11,19 @@ import be.baur.sda.Node;
 import be.baur.sda.SDA;
 import be.baur.sda.io.Parser;
 import be.baur.sda.io.SDAParseException;
-import be.baur.sds.AnyType;
+import be.baur.sds.AnyNodeType;
 import be.baur.sds.Component;
-import be.baur.sds.DataType;
 import be.baur.sds.NodeType;
 import be.baur.sds.Schema;
+import be.baur.sds.ValueNodeType;
 import be.baur.sds.common.Interval;
 import be.baur.sds.common.NaturalInterval;
 import be.baur.sds.model.ChoiceGroup;
 import be.baur.sds.model.ModelGroup;
 import be.baur.sds.model.SequenceGroup;
 import be.baur.sds.model.UnorderedGroup;
-import be.baur.sds.types.BooleanType;
-import be.baur.sds.types.CharacterType;
-import be.baur.sds.types.ComparableType;
+import be.baur.sds.types.CharacterNodeType;
+import be.baur.sds.types.ComparableNodeType;
 
 
 /**
@@ -70,7 +69,8 @@ public final class SDSParser implements Parser<Schema> {
 	private static final String NODE_NAME_INVALID = "'%s' is not a valid node name";
 	private static final String NAME_NOT_EXPECTED = "name '%s' is not expected";
 	private static final String NAME_IS_EXPECTED = "a name is expected";
-	
+
+
 	/**
 	 * Creates a schema from a character input stream in SDS format.
 	 * 
@@ -175,10 +175,10 @@ public final class SDSParser implements Parser<Schema> {
 		boolean isNodeType = sds.getName().equals(Components.NODE.tag); // will be false for a model group
 		List<Node> complexChildren = sds.getAll(n -> ! n.isLeaf()); // list of complex children (if any)
 		
-		// Simple types and references MUSt have a content type, complex types MAY have one
+		// Simple types and references MUST have a content type, complex types MAY have one
 		DataNode type = getAttribute(sds, Attribute.TYPE, isNodeType && complexChildren.isEmpty());
-		boolean isAnyType = (type == null) ? false : type.getValue().equals(AnyType.NAME);
-		boolean isDataType = (type == null) ? false : Schema.isDataType(type.getValue());
+		boolean isAnyType = (type == null) ? false : type.getValue().equals(AnyNodeType.NAME);
+		boolean isRegType = (type == null) ? false : Schema.isRegisteredType(type.getValue());
 		
 		Component component; // the component to be returned at the end of this method
 		
@@ -190,7 +190,7 @@ public final class SDSParser implements Parser<Schema> {
 			if (isAnyType) { // an any type cannot have components or attributes (except NAME and OCCURS)
 				
 				if (! complexChildren.isEmpty())
-					throw exception(sds, ATTRIBUTE_INVALID, Attribute.TYPE.tag, AnyType.NAME, "node defines content");
+					throw exception(sds, ATTRIBUTE_INVALID, Attribute.TYPE.tag, AnyNodeType.NAME, "node defines content");
 				
 				List<Node> alist = sds.getAll(n -> n.isLeaf() && 
 					! (n.getName().equals(Attribute.OCCURS.tag) || n.getName().equals(Attribute.TYPE.tag)));
@@ -201,10 +201,10 @@ public final class SDSParser implements Parser<Schema> {
 				if (! name.isEmpty() && ! SDA.isName(name))
 					throw exception(sds, NODE_NAME_INVALID, name);
 				
-				component = new AnyType(name);
+				component = new AnyNodeType(name);
 			}
 			
-			else if (isDataType || type == null) // a known data type or complex type
+			else if (isRegType || type == null) // a known data type or complex type
 				component = parseNodeType(sds, type);
 			
 			else // component must be a type reference
@@ -371,7 +371,7 @@ public final class SDSParser implements Parser<Schema> {
 		/*
 		 * Preconditions: the caller has already verified this node has a valid tag, 
 		 * and that all attributes have valid tags as well. This method is NOT called
-		 * for an AnyType.
+		 * for an AnyNodeType.
 		 * Postcondition: the caller will set the multiplicity on the returned type.
 		 */
 
@@ -396,15 +396,15 @@ public final class SDSParser implements Parser<Schema> {
 		/*
 		 * Get an instance of the requested data type and handle remaining attributes.
 		 */
-		DataType dataType = Schema.getDataType(type.getValue(), name);
+		ValueNodeType vntype = Schema.getRegisteredType(type.getValue(), name);
 		
 		// Set the optional null-ability.
 		DataNode nullable = getAttribute(sds, Attribute.NULLABLE, false);
-		if (nullable != null) switch(nullable.getValue()) {
-			case BooleanType.TRUE : dataType.setNullable(true); break;
-			case BooleanType.FALSE : dataType.setNullable(false); break;
-			default : 
-				throw exception(nullable, ATTRIBUTE_INVALID, 
+		if (nullable != null) 
+			switch(nullable.getValue()) {
+				case "true" : vntype.setNullable(true); break;
+				case "false" : vntype.setNullable(false); break;
+				default : throw exception(nullable, ATTRIBUTE_INVALID, 
 					Attribute.NULLABLE.tag, nullable.getValue(), "must be 'true' or 'false'");
 		}
 		
@@ -412,18 +412,18 @@ public final class SDSParser implements Parser<Schema> {
 		DataNode regexp = getAttribute(sds, Attribute.PATTERN, false);
 		if ( regexp != null) 
 		try { 
-			dataType.setPattern( Pattern.compile(regexp.getValue()) ); 
+			vntype.setPattern( Pattern.compile(regexp.getValue()) ); 
 		} catch (PatternSyntaxException e) {
 			throw exception(regexp, 
 				ATTRIBUTE_INVALID, Attribute.PATTERN.tag, regexp.getValue(), e.getMessage());
 		}
 		
 		// Set the length (only allowed on character data types).
-		DataNode length = getAttribute(sds, Attribute.LENGTH, dataType instanceof CharacterType ? false : null);
+		DataNode length = getAttribute(sds, Attribute.LENGTH, vntype instanceof CharacterNodeType ? false : null);
 		if (length != null) {
 			try {
 				NaturalInterval interval = NaturalInterval.from(length.getValue());
-				((CharacterType) dataType).setLength(interval);
+				((CharacterNodeType) vntype).setLength(interval);
 			} catch (IllegalArgumentException e) {
 				throw exception(length, ATTRIBUTE_INVALID, 
 					Attribute.LENGTH.tag, length.getValue(), e.getMessage());
@@ -431,10 +431,10 @@ public final class SDSParser implements Parser<Schema> {
 		}
 		
 		// Set the value range (only allowed on comparable types)
-		DataNode range = getAttribute(sds, Attribute.VALUE, dataType instanceof ComparableType ? false : null);
+		DataNode range = getAttribute(sds, Attribute.VALUE, vntype instanceof ComparableNodeType ? false : null);
 		if (range != null) {
 			Interval interval;
-			ComparableType comparableType = (ComparableType) dataType;
+			ComparableNodeType comparableType = (ComparableNodeType) vntype;
 			try {	
 				interval = Interval.from(range.getValue(), comparableType.valueConstructor());
 			} catch (IllegalArgumentException e) {
@@ -444,7 +444,7 @@ public final class SDSParser implements Parser<Schema> {
 			comparableType.setInterval(interval);
 		}
 		
-		return dataType;
+		return vntype;
 	}