+++ src/main/java/be/baur/sds/types/StringNodeType.java
@@ -3,11 +3,11 @@ package be.baur.sds.types;
 import java.util.function.Function;
 
 /**
- * A <code>StringType</code> defines an SDA node with character data. This is
- * the most basic type of all, without restrictions on value; any string is a
- * valid string, and unlike other data types it is nullable by default.
+ * A <code>StringNodeType</code> defines an SDA node with a string value. This is
+ * the most basic type of all, without restrictions on the value; any string is a
+ * valid string, and unlike other node types it is nullable by default.
  */
-public final class StringType extends CharacterType<String> {
+public final class StringNodeType extends CharacterNodeType <String> {
 
 	/** The SDS name of this data type. */
 	public static final String NAME = "string";
@@ -19,18 +19,18 @@ public final class StringType extends CharacterType<String> {
 
 
 	/**
-	 * Creates the type with the specified name.
+	 * Creates a string node type with the specified name.
 	 * 
 	 * @param name a valid node name
 	 * @throws IllegalArgumentException if the name is invalid
 	 */
-	public StringType(String name) {
+	public StringNodeType(String name) {
 		super(name); setNullable(true);
 	}
 
 
 	@Override
-	public String getType() {
+	public String getValueType() {
 		return NAME;
 	}